Upload of a simple terminal program sample project
This Atollic TrueSTUDIO project contains the sample code which allows a STM32 micro-controller to communicate with an Xbee module.
Typical use case is one where you are running some STM32 development platform (in my case the Nucleo-F303K8) and have it connected to a PC via an USB wire. By sending commands to the STM32 micro-controller over the serial UART connection, for example by using the program PuTTY, you make it possible to directly communicate with the Xbee module that is connected to the other configured UART connection on the STM32 micro-controller. 

As of right now there is no functionality attatched to the connection between the STM32 micro-controller and the external Xbee module,
only a basic terminal program has been created thus far. More updates will come in the following weeks, stay tuned!